requirements.txt
Flask
gunicorn
requests
pytz
tiktoken
google-cloud-texttospeech
google-cloud-storage
pydub
langid
sqlalchemy
openai
numpy
nest_asyncio
pycryptodome
beautifulsoup4
line-bot-sdk
pyodbc
python-dateutil
pymysql